Remove items-info category/router to items.
