+++ api/server.js
@@ -9,7 +9,6 @@ const helmet = require('helmet')
 // ----- Router Imports -----
 const authRouter = require('./auth/auth-router')
 const itemsRouter = require('./items/items-router')
-const itemsInfoRouter = require('./items-info/items-info-router')
 const itemsImagesRouter = require('./items-images/items-images-router')
 const itemsColorsRouter = require('./items-colors/items-colors-router')
 const itemsMaterialsRouter = require('./items-materials/items-materials-router')
@@ -25,7 +24,6 @@ server.use(logger('dev'))
 
 // ------- Routers --------
 server.use('/', authRouter)
-server.use('/items-info', itemsInfoRouter) //use for dev at the moment until can make proper query pulling all data includings colors and materials and also add the image_ids? to items table?
 server.use('/items', itemsRouter)
 server.use('/images', itemsImagesRouter)
 server.use('/items-colors', itemsColorsRouter)
